move stringTools functions to a cpp file
